+++ src/resdk/resolwe.py
@@ -462,7 +462,11 @@ class Resolwe:
         return Data(resolwe=self, **model_data)
 
     def _download_files(
-        self, files: List[Union[str, Path]], download_dir=None, show_progress=True
+        self,
+        files: List[Union[str, Path]],
+        download_dir=None,
+        show_progress=True,
+        custom_file_names: Union[List[str], None] = None,
     ):
         """Download files.
 
@@ -473,6 +477,8 @@ class Resolwe:
         :type files: list of file URI
         :param download_dir: download directory
         :type download_dir: string
+        :param custom_file_names: list of file names to save the downloaded files as
+        :type custom_file_names: list of strings or None
         :rtype: None
 
         """
@@ -484,9 +490,16 @@ class Resolwe:
                 "Download directory does not exist: {}".format(download_dir)
             )
 
+        if not custom_file_names:
+            custom_file_names = len(files) * [None]
+        else:
+            if not len(files) == len(custom_file_names):
+                raise ValueError(
+                    "Number of files and their corresponding custom file names must be equal."
+                )
+
         if not files:
             self.logger.info("No files to download.")
-
         else:
             self.logger.info("Downloading files to %s:", download_dir)
             # Store the sizes of files in the given directory.
@@ -494,7 +507,7 @@ class Resolwe:
             sizes: dict[str, dict[str, int]] = defaultdict(lambda: defaultdict(int))
             checksums: dict[str, dict[str, int]] = defaultdict(lambda: defaultdict(int))
 
-            for file_uri in files:
+            for file_uri, custom_file_name in zip(files, custom_file_names):
                 file_name = os.path.basename(file_uri)
                 file_path = os.path.dirname(file_uri)
                 file_url = urljoin(self.url, "data/{}".format(file_uri))
@@ -518,12 +531,19 @@ class Resolwe:
 
                 file_size = sizes[file_directory][file_name]
 
+                if custom_file_name:
+                    desc = f"Downloading file {file_name} as {custom_file_name}"
+                    actual_file_name = custom_file_name
+                else:
+                    desc = f"Downloading file {file_name}"
+                    actual_file_name = file_name
+
                 with tqdm.tqdm(
                     total=file_size,
                     disable=not show_progress,
-                    desc=f"Downloading file {file_name}",
+                    desc=desc,
                 ) as progress_bar, open(
-                    os.path.join(download_dir, file_path, file_name), "wb"
+                    os.path.join(download_dir, file_path, actual_file_name), "wb"
                 ) as file_handle:
                     response = self.session.get(file_url, stream=True, auth=self.auth)
 
@@ -540,10 +560,12 @@ class Resolwe:
                     # checksums that are difficult to reproduce here.
                     return
                 expected_md5 = checksums[file_directory][file_name]
-                computed_md5 = md5(os.path.join(download_dir, file_path, file_name))
+                computed_md5 = md5(
+                    os.path.join(download_dir, file_path, actual_file_name)
+                )
                 if expected_md5 != computed_md5:
                     raise ValueError(
-                        f"Checksum of downloaded file {file_name} does not match the expected value."
+                        f"Checksum of downloaded file {actual_file_name} does not match the expected value."
                     )
 
     def data_usage(self, **query_params):