+++ tests/unit/test_data.py
@@ -178,13 +178,17 @@ class TestData(unittest.TestCase):
 
         Data.download(data_mock)
         data_mock.resolwe._download_files.assert_called_once_with(
-            ["123/file1.txt", "123/file2.fq.gz"], None
+            files=["123/file1.txt", "123/file2.fq.gz"],
+            download_dir=None,
+            custom_file_names=None,
         )
 
         data_mock.reset_mock()
         Data.download(data_mock, download_dir="/some/path/")
         data_mock.resolwe._download_files.assert_called_once_with(
-            ["123/file1.txt", "123/file2.fq.gz"], "/some/path/"
+            files=["123/file1.txt", "123/file2.fq.gz"],
+            download_dir="/some/path/",
+            custom_file_names=None,
         )
 
     @patch("resdk.resolwe.Resolwe")